改善

+++ server.py
@@ -5,13 +5,13 @@ app =Flask(__name__)
 date = datetime.datetime.now().strftime('%Y%m%d')
 file_path="./sensor_data_" + date +".csv"
 port_num=21118
-#if os.path.isfile(file_path)==False :
-f = open(file_path, 'w')
-f.write("time"+"," + "value" + "\n")
-f.close()
+if os.path.isfile(file_path)==False :
+  f = open(file_path, 'w')
+  f.write("time"+"," + "value" + "\n")
+  f.close()
 @app.route('/',methods=['GET'])
 def get_html():
-    return render_template('./index2.html')
+    return render_template('./index.html')
 @app.route('/lux',methods=['POST'])
 def update_lux():
    time = request.form["time"]
